+++ envs/monkey_zoo/blackbox/utils/config_generation_script.py
@@ -3,7 +3,6 @@ import pathlib
 from typing import Type
 
 from envs.monkey_zoo.blackbox.config_templates.config_template import ConfigTemplate
-from envs.monkey_zoo.blackbox.config_templates.drupal import Drupal
 from envs.monkey_zoo.blackbox.config_templates.hadoop import Hadoop
 from envs.monkey_zoo.blackbox.config_templates.log4j_logstash import Log4jLogstash
 from envs.monkey_zoo.blackbox.config_templates.log4j_solr import Log4jSolr
@@ -49,7 +48,6 @@ CONFIG_TEMPLATES = [
     WmiMimikatz,
     WmiPth,
     Zerologon,
-    Drupal,
     Log4jLogstash,
     Log4jTomcat,
     Log4jSolr,