+++ monkey/monkey_island/cc/services/config_schema/basic.py
@@ -21,7 +21,6 @@ BASIC = {
                         "WebLogicExploiter",
                         "HadoopExploiter",
                         "MSSQLExploiter",
-                        "DrupalExploiter",
                         "PowerShellExploiter",
                     ],
                 }